requirements.txt
streamlit
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
numpy==2.0.1
pandas
plotly
